monitoring/exporters/vertex-ai-exporter/requirements.txt
google-cloud-aiplatform==1.38.1
prometheus-client==0.19.0
flask==2.3.3
gunicorn==21.2.0
